_uois-latest/gql_forms/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]
uvicorn[standard]
gunicorn

asyncpg
sqlalchemy-cockroachdb

aiohttp
requests

sqlalchemy
sqlalchemy_utils

pydantic
starlette
strawberry-graphql
click
aiodataloader
pyjwt[crypto]

prometheus_fastapi_instrumentator
https://github.com/hrbolek/uoishelpers/archive/refs/heads/main.zip

opentelemetry-sdk 
opentelemetry-instrumentation-fastapi 
opentelemetry-exporter-otlp

